started teacher availability and flux usage done
