+++ package.json
@@ -6,12 +6,14 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "bootstrap": "^5.1.1",
+    "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.5.0",
     "bootswatch": "^5.1.2",
     "jquery": "^3.6.0",
+    "moment": "^2.29.1",
     "popper.js": "^1.16.1",
     "react": "^17.0.2",
+    "react-availability-calendar": "^0.3.13",
     "react-bootstrap": "^2.0.0-rc.0",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.3.0",
